requirements.txt
Django==5.1.5
python-dotenv==1.0.1
bcrypt==4.2.1
psycopg2[binary]
django-filter==24.3
djangorestframework==3.15.2
markdown==3.7
djangorestframework-simplejwt==5.4.0
django-cors-headers == 4.6.0
django-extensions==3.2.3
pytz==2025.1
